+++ pi-web-apis/traffic-api/README.md
@@ -1,0 +1,3 @@
+This has not been implemented yet. However, we have a placeholder payload that 
+uses modbus to change the traffic light. Curl and run this: 
+https://hst.sh/raw/qayoniside
